feat: ini root index
