--- packages/modals/src/deposit/deposit-modal.stories.tsx
@@ -1,69 +1,0 @@
-// Copyright 2023-2024 Light, Inc.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//   http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-/* eslint-disable react-hooks/rules-of-hooks */
-
-import { useAuth, useModals } from "@lightdotso/stores";
-import type { Meta, StoryObj } from "@storybook/react";
-import { useEffect } from "react";
-import TokenModal from "../token/token-modal";
-import { DepositModal } from "./deposit-modal";
-
-// -----------------------------------------------------------------------------
-// Meta
-// -----------------------------------------------------------------------------
-
-const meta: Meta<typeof DepositModal> = {
-  title: "modal/DepositModal",
-  component: DepositModal,
-  tags: ["autodocs"],
-  argTypes: {},
-};
-
-export default meta;
-
-// -----------------------------------------------------------------------------
-// Types
-// -----------------------------------------------------------------------------
-
-type Story = StoryObj<typeof DepositModal>;
-
-// -----------------------------------------------------------------------------
-// Story
-// -----------------------------------------------------------------------------
-
-export const Base: Story = {
-  render: () => {
-    const { setAddress } = useAuth();
-    const { showDepositModal } = useModals();
-
-    useEffect(() => {
-      setAddress("0x4fd9D0eE6D6564E80A9Ee00c0163fC952d0A45Ed");
-      // eslint-disable-next-line react-hooks/exhaustive-deps
-    }, []);
-
-    useEffect(() => {
-      showDepositModal();
-      // eslint-disable-next-line react-hooks/exhaustive-deps
-    }, []);
-
-    return (
-      <>
-        <DepositModal />
-        <TokenModal />
-      </>
-    );
-  },
-  args: {},
-};
